+++ app/src/main/java/com/example/traders/di/SharedPrefsModule.kt
@@ -1,11 +1,10 @@
-package com.example.traders.hilt
+package com.example.traders.di
 
 import android.content.Context
 import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ActivityContext
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
