+++ app/src/main/java/com/example/traders/di/WebSocketModule.kt
@@ -1,4 +1,4 @@
-package com.example.traders.hilt
+package com.example.traders.di
 
 import com.example.traders.webSocket.BinanceWSClient
 import com.example.traders.webSocket.BinanceWSClientImpl
